requirements.txt
Flask==2.1.1
openai==0.27.0
tweepy==4.1.0
Pillow==9.0.1
python-dotenv==0.19.1
